OneDrive/Desktop/Yotube_summarizer/requirements.txt
flask
yt-dlp
anthropic
python-dotenv
gunicorn
